bi_ce_container/requirements.txt
Flask==3.0.2
huggingface_hub==0.20.3
langchain==0.1.9
numpy==1.26.4
packaging==23.2
pandas==2.2.1
PyYAML==6.0.1
PyYAML==6.0.1
Requests==2.31.0
sentence_transformers==2.3.0
torch==2.2.0
tqdm==4.66.1
transformers==4.36.2
